Move envelope processor to handler package
